requirements.txt
# Flask core
flask==3.0.2
werkzeug==3.0.1
jinja2==3.1.3

# Excel
pandas==2.2.1
openpyxl==3.1.2
xlsxwriter==3.1.9
numpy==1.26.4

# Word (Estudios Previos)
python-docx==1.1.2

# Scraping
playwright==1.42.0
requests==2.32.3

# Servidor de producción
gunicorn==21.2.0
